Added Tests:
There seems to be some sort of memory leak/sigfault going on for the ConstantTimeDynamicArray when using uneven length. There only is a sigfault in totally different tests following the tests for the array.
The Array tests all run through, the destructors are called correctly and the error makes no sense to me.